requirements.txt
# python packages needed
openai
langchain
streamlit
chainlit
pypdf3
python-dotenv
faiss-cpu
huggingface_hub



